+++ Source/SwiftLintFramework/Models/ConfigurationError.swift
@@ -2,4 +2,6 @@
 public enum ConfigurationError: Error {
     /// The configuration didn't match internal expectations.
     case unknownConfiguration
+    /// The configuration had both `match_kind` and `excluded_match_kind` parameters.
+    case ambiguousMatchKindParameters
 }