+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 * Improves description for  `empty_enum_arguments` 
   [Lukas Schmidt](https://github.com/lightsprint09)
 
+* Add support for `excluded_match_kinds` custom rule config parameter.  
+  [Ryan Demo](https://github.com/ryandemo)
+
 #### Bug Fixes
 
 * None.